Update tool definitions to support expected_hash and version
Previously it was impossible to have multiple versions of the same
tool because tools were indexed by name. This made it hard to upgrade
versions automatically because the existing tool was considered to
satisfy the dependency.

This PR adds a version field to the tool definitions making the
combination of name and version unique. This also allows an
artifact to declare a specific version of a tool to use.

Additionally the tool definition now supports an expected hash field
which declares what the hash the artifact writer expects for the
tool. Velociraptor will download the tool and ensure the hash is the
same as the expected hash to avoid the tool being changed upstream.