Extend http_client() to support SMB urls. (#2635)
This PR abstracts all HTTP client objects to a special wrapper with
support for additional schemes. Mainly this supports the file:// and
smb:// scheme.

This allows an SMB connection to be made whenever a URL is expected. For
example when serving tools to clients it is possible to provide a URL of
the form:

smb://test:testpasswd@192.168.1.1/Tools/tool.exe

This will cause the clients to connect to the SMB share for downloading
the tools instead of the Velociraptor server.