Allow SQLite files to be copied always.
When the SQLite file is in use, the SQLite driver takes about 10
seconds to retry opening the file. This slows down processing a
lot. This PR just copied the file to a temp file for every valid
SQLite file because it is much faster than attempting to open the
original.