Bugfix: Sort flows before fetching them into the GUI
Currently the GUI only shows the first 1000 flows in each client. If
more than that many flows are available, the most recent flows will
sometimes be truncated which makes the GUI not show them. This looks
to the user as if the flow did not launch but it actually did.

This PR also reduced the number of flows to show in the GUI at the 200
most recent flows.