Implemented memory protections for notebook cell calculations (#2842)
These help to protect the server when a notebook query uses too much
memory. When a new notebook query is started, the notebook checks if the
current memory use is smaller than the low water mark, if not, the
notebook will wait for a while until this is ok.

Once started the notebook will run a nanny to ensure the memory does not
exceed the high water mark. If the memory exceeds the high water mark,
the notebook will abort the query with insufficient resources.