Support deb source option add/remove
I missed a feature of both the one-line and deb822 formats, allowing one
to add or remove values from the default without overring the defaults.
The syntax is "[lang-=en]" in one-line and "Languages-Remove: en" in
deb822.

I doubt that these are used often, but I spotted a use in the wild, so
it should be supported.