Add simple support for quoted values
In particular, parse "cdrom" sources lines correctly. These are not
uncommon in most OSes after an install, albeit typically commented out.

Quoting URIs, suites and components seems to be supported in the apt
source code, but processing these fails later in apt. Support is not
documented. However, the use of [] as quotes is often used for the
"cdrom" protocol. Modify the regex so that these lines are parsed
correctly. Also support fully quoted URIs in "", but stop there.

Also support comments at the end of options, as the source code mentions
that this is supported (this is an easy adjustment of the regex).