Bugfix: accessors should provide their underlying file
We sometimes need to know the path to the underlying file on the
filesystem if possible. This is used in cases when we need to delegate
to an external library which expects a filesystem path.

Previously the code assumed that when the accessor was "file" or
"auto" then the underlying path can be obtained from the
filename. This works well in local trigage mode but fails when
remapping - in that case the actual accessor called "file" may be a
completely different remapped accessor and it is not appropriate to
use its filename as an underlying API file.

This would cause issues with e.g. the yara plugin, sqlite and leveldb
plugins.

This PR introduces a new interface which allows the accessor to
provide the raw API accessible path if possible. For plugins that need
to work with real files, this path also creates a local copy if
needed.

Fixes: #2870