Bugfix: Allow serve url to be set without materializing
Some servers have locked down egress and can not materialize the tool
themselves (to calculate the tool hash). This means that the user is
not able to set an alternative serve url from these servers without
triggering a materialize (which fails)

This PR does not require the server to materialize the tool when
updating the serve url. This allows a user to manually upload a
tool (thereby calculating its hash) then change the serve url,
preserving the hash - all without the server fetching the tool from
the internet.