Allow embedded config to come from an external file (#2899)
On MacOS we can not modify the binary after build due to hashes and
verifications. This PR allows the embedded config to be written to a
stub file (shell script) which can be used later to bootstrap the
offline collector.

This approach is a bit more complicated but will also be useful on
Windows where the repack operation invalidates the signature. We can
still use this same approach on Windows too without needing to resign
the binary.