Disable compatibility with URL style paths
In the distant past we used to use url encoding to encode recursive
pathspecs but this is problematic because url encoding is just not
reliable enough. We have since moved to an OSPath object that properly
encodes paths as JSON objects but there is still some backwards
compatibility code in there that tries to detect a URL style path by
looking for # (the fragment component of the url).

This is probably a good time to remove this compatibility code because
we dont use urls for probably about 4-5 releases now.

Fixes: #2906