Relay hunt creation errors to the Hunts API (#2953)
During refactor of the CreateHunt API to be able to support multiple
orgs the errors from launching the hunt on any org were reported in the
log and not relayed to the caller.

This means that the errors were not propagated to the API so if the
artifact failed to compile within the specified org no error was visible
to the user.

This PR collects all errors and reports them together therefore ensuring
the user knows the hunt can not be launched.

FIXES: #2952