Initial implementation of native Sigma support (#3020)
This PR introduces two plugins:

* The sigma_log_source() function prepare log sources which are a
mapping between the sigma log source and a query that will deliver
events for that log source.

* The sigma() plugin accepts a set of sigma rules and a log source and
applies the rules to the log sources.

Only the relevant log sources will be used. Log sources which have no
rules against them will be ignored. This allow the query to skip parsing
event logs which are not relevant.

Field mappings are translations between the event fields and the fields
as referred in the sigma rule. In this plugin field mappings are
expressed as VQL lambdas. This PR includes all the standard ones. It is
possible to create non standard fields by calling arbitrary code in the
field mapping.