Allow minion workers to have different priority.
Minion workers can now specify a higher (or lower) priority which
means that notebook queries will be directed to the minions rather
than running on the master node.

Minion preferences are now set in the config file in a Minion section:

```yaml
Minion:
   notebook_worker_priority: 10
   notebook_number_of_local_workers: 2
```

Allows 2 workers with priority 10 to be running on each minion.