CSV Form widget can now flip between textarea or table
While the table view of the CSV artifact parameter type is intutive to
use for adding a few rows manually it is not easy to use when we have
a large collection of CSV items.

This PR adds a button to the widget to allow it to toggle to a simple
textarea. In that case the user can paste a large CSV file into the
GUI.

The GUI will now also validate the CSV warning about errors like
unbalanced columns etc.