Parse Linux group IDs, names and memberships (#3038)
I've had the need to look up group names from GIDs, as well as finding
group memberships. This simple artifact does that. It is very much like
Linux.Sys.Users, except that GID is an int. I felt that this made more
sense, since gid in stat() is an int.