Adding a local encrypted storage for clients.
This allows clients to store data securely on the endpoint without the
risk of the data being readable by anything other than the server.

Also added VQL support to be able to read/write such files.

Added the logging() plugin to allow a query to follow the logs
generated by the client.