Import hunts using the ImportCollection plugin (#3075)
Added the capability to import hunts using the import_collection plugin.
I have also modified the function to auto-detect certain aspects, such
as whether the object is a hunt or an offline collector. This will now
allow for bulk ingestion as certain required arguments for
import_collection have been removed.

This also includes adding a "hunt_info.json" file that will be created
when downloading a hunt. Because this file has all the necessary
information for importing a hunt, hunts will not be able to be imported
from before this change.

---------

Co-authored-by: Mike Cohen <mike@velocidex.com>