Allow clients to present client certificates for mTLS authentication (#3133)
This PR allows clients to include cilent side certificates and adds a
server option to require those. This additonal authentication
requirement between clients and server ensure that only clients with a
valid issued certificate can connect to the Velociraptor server.

Also added docs for ECS and a notebook to help upload to Elastic.