Mark inherited artifacts in child orgs. (#3156)
Because inherited artifacts are visible in the child org and can be
customized they were showing as user custom artifacts. The GUI allowed
the user to attempt to delete them but this did not work because the
artifact only exists in the root org.

This PR clearly marks inherited artifacts and prevents the user from
deleting them.