Bugfix: Store the repository in the scope for use in VQL
This fixes #3299 in which the VQL running on the client was accessing
the local client side repository instead of the artifacts that were
being sent by the VQL request.

This made it impossible to automatically package custom artifacts in
the vql request, although built in artifacts would work from the local
client artifact repository.

This changes forces the collect() plugin to use the query side
repository which is empty other than those artifacts explicitely
sent. This means that collect() will not be able to find built in
artifacts any more and everything needs to be sent.

```
name: Custom.CollectTest
sources:
  - query: |
      LET _ = SELECT * FROM Artifact.Custom.Artifact()
      SELECT * FROM collect(artifacts="Custom.Artifact")
```