Added a COLLECT_BASIC permission for artifacts
Currently users can receive the COLLECT_CLIENT permission which allows
them to collect any artifacts on the client.

One commonly requested permission model is to have finer grained
control over which artifacts are allowed to be collected by certain
users. This is needed when a set of low privilege users are needed to
run low risk, routine artifacts but we dont want to give them the
COLLECT_CLIENT permission which allows all artifacts to be run.

This PR introduces another permission COLLECT_BASIC. Additionally,
this PR introduces a new "basic" metadata field on an artifact (can be
set using the artifact_set_metadata() function). When a user has this
permission, they can only run artifacts that were marked as "basic".

This allows server administrators to only allow some artifacts to be
run by users who have the COLLECT_BASIC permission.