Refactored notebook storage to have in memory index (#3322)
Notebook sharing was implemented using an on disk index keyed by the
username. This resulted in being unable to detect public notebooks when
these notebooks were not directly shared with the user as well.

This PR refactores the notebook manager storage to maintain an in memory
cache of all global notebooks. This should make listing and viewing
global notebooks faster as well as removing the need for notebook share
index completely.