Bugfix: Handle directories uploaded into zip file. (#3327)
A Zip archive only contains files and has no concept of directories. If
the upload() VQL function is accidentally called on a directory,
Velociraptor will try to read the directory and store it in the zip
file. This leads to a 0 length file placed in the middle of a directory
path:

/Foo/Bar <- zero length file
/Foo/Bar/Baz <- a real file.

This makes the real file impossible to extract and third party zip
programs will be unable to write it once they created a file at the
directory level above. This also confuses Velociraptor's zip accessor
which emulates directories based on path prefixes and see /Foo/Bar as a
file not a directory (so it does not recurse into it).

This PR ensures that when a directory is passed to upload() it will have
a trailing / appended. This seems to be a convention with zip programs
to denote a directory in a zip file. Additionally we consider a zero
length file in the zip archive to be a directory regardless if it has a
trailing / in order to be able to recover older collections which were
written without the trailing /.