Sigma: Support selection matchers with null values.
The following rule matcher is valid:

```
detection:
   selection:
    field: null
```

Also support rules which omit the condition - while not strictly
correct they mean match all the selection clauses.

This PR also removes the coersion to string of all the rule values and
use scope protocols instead. This results in a significant speedup as
well.