Sigma: Support selection matchers with null values. (#3356)
The following rule matcher is valid:

```
detection:
   selection:
    field: null
```


https://github.com/SigmaHQ/sigma-specification/blob/main/Sigma_specification.md#special-field-values

Also support rules which omit the condition - while not strictly correct
they mean to match all the selection clauses.

This PR also removes the coercion to string of all the rule values and
use scope protocols instead. This results in a significant speedup as
well.