Bugfix: Crash with pipe plugin
When pipe() was used with parse_lines() a crash occured. This happened
because parse_lines() attempts to open files as gzip files first, then
reopen them again if that fails.

It is not possible to safely reopen an pipe though. This PR fixes the
crash and avoids opening the pipe as gzip first.