Added Windows.Registry.Hunter
Also added multichoice parameter types.

CreateInitialNotebook will now block until all the cells are
constructed to ensure that the cells can not overwhelm the notebook
workers. When an artifact creates a large default notebook, the
browser will attempt to render each in parallel and this will initiate
a worker update.