Separate tab org id from traits.
Add the org id to the URL so that URLs also point at the correct
org.

This PR also allows different tabs to view different orgs at the same
time.

The org stored in the user traits is now only used for the initial org
switch when the main page is viewed, then the org is encoded in the
URL and is specific for that tab.