Bugfix: Only update notebooks atomically
Notebook updates were allowed to be done without locking the
notebook. This resulted in the notebook being corrupted when updates
were initiated during times where a cell was calculated, resulting in
cell corruption.

This PR ensures the GUI does not attempt to edit the notebook while
its state is not in sync with the server. As well as that the notebook
manager ensures the notebook is updated under locks.

Also fixed memory leak regression with group by