Enable local client encrypted logs in config file.
Previously we can write local log files with the
Generic.Client.LocalLogs client artifact but this did not help to
disgnose errors in startup before the client was able to connect to
the server to receive this artifact.

This PR adds local logging as a built in capability at startup.