Support automatic user creation with SAML authentication
With this change we are able to automatically create users for SAML
authenticated users and assign them roles.
Before this change, users had to be created manually on the velociraptor
server, even when using SAML authentication, which lead to double book
keeping.

Based on the work in the certs authenticator module:
- https://github.com/Velocidex/velociraptor/blob/858e2e4a1c597c9dfd1d0389f84ac2186eb3bae7/api/authenticators/certs.go#L168-L223

Instead of reusing the gui.authenticator.default_roles_for_unknown_users
field we introduced a new saml_user_roles field to ensure privilege
separation between gui users and client certificates.

Co-Authored-By: Deniz Adrian <adrian@hisolutions.com>