Add an auditd-based process tracker event artifact for Linux (#3546)
This artifact provides updates to the process tracker in a very similar
fashion as *Windows.Events.TrackProcesses*. It uses the integrated
`audit()` function and inserts rules as it runs, and removes them in an
`atexit()` call. I haven't found a way to audit process exits when
neither "exit" nor "exit_group" are called by the process. Auditing the
sys. calls "sigreturn"/"rt_sigreturn" does not seem to help.

I haven't created any test code, as this would probably require mocking
code in go. I'll give it a shot if needed.