Bugfix: Memory leak in diff plugin
The diff plugin did not run each iteration in an isolated scope so
cleanup destructors where not being called.

Additionally registry cache was not installed in the root scope which
limited its usefulness in sub queries mostly leading to cache misses.

The default behaviour of TTL cache is to extend the ttl on each Get()
operation. This is hardly ever what we want because then some cached
items will never expire.