Special case offline collector copy collection.
The Offline Collector builder is a special GUI component used to build
an offline collector. It creates an instance of the
Server.Utils.CreateCollector artifact.

Previously when copying this collection, the default copy artifact
widget was used which allowed the user to modify the raw artifact
parameters to that artifact but this is not user friendly enough.

This PR opens the custom GUI component for editing allowing
modification of the offline collector in a more natural way.