Bugfix: Report access denied to org errors. (#3596)
Previously the org was just automatically switched to another valid org
when a user had no access to that org. This caused a problem when the
user was an org admin so could switch to any org, but didnt have
permission in the new org. This caused the app to switch to another org
but the org selector still had the old org shown.

This resulted in a confusing UI when the org selector was showing an org
but the application was actually communicating with a different org.

This change propagates the access denied error to ensure the user is
aware they can not connect to the org they wanted rather than
automatically switch to another org.

Also fixed a GUI bug where fractional timestamps were not properly
shown.