Include snaps in Linux.Debian.Packages artifact (#3612)
Like it or not, the Canonical app store for Linux, snap, is included in
Ubuntu (and possibly other Ubuntu-flavoured distros), and it can also be
installed on other Debian systems. When looking for installed packages
on a Debian system, it makes sense to also look for installed snaps. A
number of software packages are distributed in both snap and deb
formats.

The changes to the Linux.Debian.Packages artifact are summarised as
follows:

- Add a new parameter for the snapd UNIX socket
- Add two new queries, and keep the original query mostly intact, now
named "DebPackages". Changes to original query:
  - Use `humanize()` on "InstalledSize" instead of displaying bytes
- Include package description as a new column, hidden by default:
\_Description
- The new query "Snaps" lists all installed snaps, queried from the snap
daemon (snapd) using a HTTP GET request to a UNIX socket. This is by far
the easiest way to parse installed package details, avoiding having do
deal with files in compressed SquashFS filesystems. If the snapd daemon
does not respond, or if snapd is not installed, this query simply
returns nothing.
- Add a third query that combines the deb and snap package lists, with a
column "Type" that indicates the package type ("deb", "snap")

The golden test output has been updated, an example snapd JSON response
is included

---------

Co-authored-by: Mike Cohen <mike@velocidex.com>