Bugfix: Reset column toggles when changing tables (#3614)
When the user navigates to a different table the toggles set was
retained in some cases causing it to ignore the calculating of default
settings. This caused normally hidden columns to show up in some cases.

Also: Added the `copy_env` parameter to the `query()` plugin to allow
functions to be carried into the query scope.