Add artifact for installing and configuring deb packages (#3626)
In Debian-based OSes software is preferably installed through the
package manager. This artifact provides a comfortable way to install
packages, as well as configuring them using debconf. The package can
either be a name or a deb file, which can be optionally uploaded.

The purpose is to provide a reusable component for installing and
configuring tools and other packages needed by other artifacts in a
single command. `apt-get update` is optionally run, and dependencies, if
any, is automatically installed. debconf may be pre-seeded with answers
if the package requires configuration.

Other artifacts can offload all installation and configuration logic to
this artifact by calling it like this: `SELECT * FROM
Artifact.Linux.Utils.InstallDeb(DebName='sl')` or `SELECT * FROM
Artifact.Linux.Utils.InstallDeb(DebName='/tmp/mypackage_0.1.0-1_amd64.deb')`.
A row with *Step* "Installing package" and *ReturnCode* 0 will indicate
that the installation was a success.