Bugfix: Rewrite vite generated css files to honor the base path (#3668)
Vite rewrites css url() without regard to base path - this means fonts
were broken when base path was changed to serve from a deeper directory.

Also rename config's client version to server_version to clarify that is
the version of the server which generated the config file. Also we dont
need to propagate all the fields to the client's config.