Added a user_options() VQL function
This allows to update the user's GUI options like theme, lang etc
using VQL notebooks.

It is now possible for admins to configure a default theme and default
password for all users by watching new user creations and updating the
GUI options as soon as a new user is created.