Only use tempfiles when necessary in queue manager. (#3689)
Previously we opened a temp file for each queue but there are many
queues now and it is a bit unsightly. This PR only uses a tempfile only
when it is necessary (i.e. the in memory queue is full and we need to
fall back to disk backed buffer file).

Also refactored the timeline viewer to use better tables and column
selector.